refactor: move pinPackageVersion into @doenet/standalone
It lived in `@doenet/doenetml`, which meant a new public export on a
published package and a home in `global-config.ts` — a module about
`doenetGlobalConfig`, which this function has nothing to do with — chosen
because a new file under `packages/doenetml/src` makes Rollup name the big
shared chunk after it.

`@doenet/standalone` is its only consumer, and the one follow-up that
looked like a second one is not: where the iframe wrapper's shared-core
worker URL can still float, the host named the range and the parent has no
exact version to pin to, which is why that fix has to be the child
reporting the version it loaded.

So the function moves to the package that uses it, with its tests, and
`setExternalCoreWorkerUrl` stays behind — it writes `doenetGlobalConfig`
and owns the bootstrap Blob's lifetime.
