+++ packages/doenetml/src/index.ts
@@ -32,11 +32,6 @@ export type {
 
 export { CodeMirror } from "@doenet/codemirror";
 
-// For a bundle served from a CDN under a floating tag: pins the URLs it
-// resolves its sibling assets against — the core worker, the message catalogs —
-// to the release the bundle itself is, so caches cannot pair it with another's.
-export { pinPackageVersion } from "./global-config";
-
 // Where a host installs the message catalogs the viewer in *this* bundle reads.
 //
 // The loaders are module-level state, and a bundle can hold more than one
